+++ app/graphql/resolver-types.ts
@@ -95,7 +95,6 @@ export type DataCubeMeasuresArgs = {
 
 export type DataCubeComponentFilter = {
   iri: Scalars['String'];
-  latest?: Maybe<Scalars['Boolean']>;
   filters?: Maybe<Scalars['Filters']>;
   componentIris?: Maybe<Array<Scalars['String']>>;
   joinBy?: Maybe<Scalars['String']>;
@@ -106,12 +105,10 @@ export type DataCubeComponentFilter = {
 
 export type DataCubeMetadataFilter = {
   iri: Scalars['String'];
-  latest?: Maybe<Scalars['Boolean']>;
 };
 
 export type DataCubeObservationFilter = {
   iri: Scalars['String'];
-  latest?: Maybe<Scalars['Boolean']>;
   filters?: Maybe<Scalars['Filters']>;
   componentIris?: Maybe<Array<Scalars['String']>>;
   joinBy?: Maybe<Scalars['String']>;
@@ -127,7 +124,6 @@ export type DataCubeOrganization = {
 
 export type DataCubePreviewFilter = {
   iri: Scalars['String'];
-  latest?: Maybe<Scalars['Boolean']>;
 };
 
 export enum DataCubePublicationStatus {
@@ -137,7 +133,6 @@ export enum DataCubePublicationStatus {
 
 export type DataCubeTermsetFilter = {
   iri: Scalars['String'];
-  latest?: Maybe<Scalars['Boolean']>;
 };
 
 export type DataCubeTheme = {
@@ -333,6 +328,7 @@ export type OrdinalMeasureValuesArgs = {
 
 export type Query = {
   __typename?: 'Query';
+  dataCubeLatestIri: Scalars['String'];
   dataCubeComponents: Scalars['DataCubeComponents'];
   dataCubeComponentTermsets: Array<Scalars['ComponentTermsets']>;
   dataCubeMetadata: Scalars['DataCubeMetadata'];
@@ -341,6 +337,13 @@ export type Query = {
   possibleFilters: Array<ObservationFilter>;
   searchCubes: Array<SearchCubeResult>;
   dataCubeDimensionGeoShapes?: Maybe<Scalars['GeoShapes']>;
+};
+
+
+export type QueryDataCubeLatestIriArgs = {
+  iri: Scalars['String'];
+  sourceType: Scalars['String'];
+  sourceUrl: Scalars['String'];
 };
 
 
@@ -941,6 +944,7 @@ export type OrdinalMeasureResolvers<ContextType = VisualizeGraphQLContext, Paren
 }>;
 
 export type QueryResolvers<ContextType = VisualizeGraphQLContext, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = ResolversObject<{
+  dataCubeLatestIri?: Resolver<ResolversTypes['String'], ParentType, ContextType, RequireFields<QueryDataCubeLatestIriArgs, 'iri' | 'sourceType' | 'sourceUrl'>>;
   dataCubeComponents?: Resolver<ResolversTypes['DataCubeComponents'], ParentType, ContextType, RequireFields<QueryDataCubeComponentsArgs, 'sourceType' | 'sourceUrl' | 'locale' | 'cubeFilter'>>;
   dataCubeComponentTermsets?: Resolver<Array<ResolversTypes['ComponentTermsets']>, ParentType, ContextType, RequireFields<QueryDataCubeComponentTermsetsArgs, 'sourceType' | 'sourceUrl' | 'locale' | 'cubeFilter'>>;
   dataCubeMetadata?: Resolver<ResolversTypes['DataCubeMetadata'], ParentType, ContextType, RequireFields<QueryDataCubeMetadataArgs, 'sourceType' | 'sourceUrl' | 'locale' | 'cubeFilter'>>;